loginsystem/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==24.2
django-six==1.0.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
importlib_metadata==7.1.0
isoweek==1.3.3
Markdown==3.6
PyJWT==2.8.0
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.12.1
tzlocal==5.2
xlwt==1.3.0
zipp==3.19.2
psycopg2-binary>=2.9.1